change resolve and signature to use new function
