+++ app/js/features/arbitration/actions.js
@@ -1,5 +1,18 @@
-import { GET_DISPUTED_ESCROWS, RESOLVE_DISPUTE } from './constants';
+import {ARBITRATION_UNSOLVED, GET_DISPUTED_ESCROWS, RESOLVE_DISPUTE, RESOLVE_DISPUTE_FAILED} from './constants';
+import Escrow from 'Embark/contracts/Escrow';
 
-export const getDisputedEscrows = () => ({ type: GET_DISPUTED_ESCROWS });
+export const getDisputedEscrows = () => ({type: GET_DISPUTED_ESCROWS});
 
-export const resolveDispute = (escrowId, result) => ({ type: RESOLVE_DISPUTE, escrowId, result: result.toString() });
+export const resolveDispute = (escrowId, result) => {
+  if (result === ARBITRATION_UNSOLVED) {
+    return {
+      type: RESOLVE_DISPUTE_FAILED,
+      result: "Arbitration must have a result"
+    };
+  }
+  return {
+    type: RESOLVE_DISPUTE,
+    escrowId,
+    toSend: Escrow.methods.setArbitrationResult(escrowId, result)
+  };
+};
